+++ ft_strnstr.c
@@ -6,17 +6,17 @@
 /*   By: juramos <juramos@student.42madrid.com>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/18 10:49:40 by juramos           #+#    #+#             */
-/*   Updated: 2023/09/18 11:13:45 by juramos          ###   ########.fr       */
+/*   Updated: 2023/09/18 13:05:42 by juramos          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-char	*ft_strnstr(char *haystack, char *needle, int len)
+char	*ft_strnstr(char *haystack, char *needle, size_t len)
 {
-	int	h_pos;
-	int	n_pos;
-	int	needle_length;
+	size_t	h_pos;
+	size_t	n_pos;
+	size_t	needle_length;
 
 	h_pos = 0;
 	needle_length = ft_strlen(needle);
